requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database ORM
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.0.0

# Templates + static files
jinja2>=3.1.3
aiofiles>=23.2.1

# Environment variables
python-dotenv>=1.0.0

# Needed for form data parsing (FastAPI)
python-multipart>=0.0.9
